requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.0
requests>=2.31.0
deep-translator>=1.11.0
gTTS>=2.4.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
PyMuPDF>=1.23.0
Pillow>=10.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
